Petites corrections mineures avant le tag

+++ ASService.m
@@ -182,7 +182,7 @@
  * Max allowed length of user names of this service.  Account and contact names longer than this will not be allowed.
  * @return Max name length
  */
-- (int)allowedLength
+- (NSUInteger)allowedLength
 {
     return 100;
 }
